[Feat] 그룹 이미지 저장, 수정, 조회 로직 추가
