+++ src/pages/band/LeftBarBandList.vue
@@ -2,15 +2,17 @@
 export default {
   data() {
     return {
+      leftBarLoading: false,
       myBandList: [],
+      myBandImageMap: new Map(),
     };
   },
   created() {
     this.fetchData();
   },
   methods: {
-    fetchData() {
-      this.$axios
+    async fetchData() {
+      await this.$axios
         .get("/user/api/v1/users/my-bands")
         .then((response) => {
           this.myBandList = response.data;
@@ -18,6 +20,23 @@ export default {
         .catch(() => {
           alert("left var server error");
         });
+
+      for (const myBand of this.myBandList) {
+        if (myBand.fileNo !== null) {
+          const response = await this.$axios
+            .get("/file/api/v1/files/" + myBand.fileNo, {
+              headers: {
+                Accept: "application/json",
+              },
+            })
+            .catch(() => {
+              alert("이미지 로드에 실패하였습니다.");
+            });
+
+          this.myBandImageMap.set(myBand.bandNo, response.data.path);
+        }
+      }
+      this.leftBarLoading = true;
     },
   },
 };
@@ -28,7 +47,7 @@ export default {
     <label class="custom-my-band-title">내 그룹</label>
     <v-divider class="custom-divider"></v-divider>
     <v-divider class="custom-divider"></v-divider>
-    <div class="band-list-flex">
+    <div v-if="leftBarLoading" class="band-list-flex">
       <router-link
         class="remove-line d-inline-block"
         v-for="myBand in myBandList"
@@ -42,12 +61,17 @@ export default {
                 class="image-container"
                 alt="no image"
                 max-width="40px"
-                src="../../assets/band-no-image.png"
+                :src="
+                  myBandImageMap.has(myBand.bandNo)
+                    ? myBandImageMap.get(myBand.bandNo)
+                    : require('@/assets/band-no-image.png')
+                "
               >
               </v-img>
             </span>
           </template>
           <span> {{ myBand.name }}</span>
+          <span> {{ myBand.bandNo }}</span>
         </v-tooltip>
       </router-link>
     </div>
@@ -63,6 +87,7 @@ export default {
 }
 
 .image-container {
+  aspect-ratio: 1/1;
   margin-left: 0.4em;
   border-radius: 10px;
   border: 2px solid darkgray;
